+++ module/actes-generique/action/ActesGeneriqueArreteChange.class.php
@@ -2,6 +2,7 @@
 class ActesGeneriqueArreteChange extends ActionExecutor{
 
 	public function go(){
+		
 		$content_type = $this->getDonneesFormulaire()->getContentType('arrete');
 		//Vérifier que le doc est en xml ou en pdf
 		if (in_array($content_type,array("application/pdf","application/xml"))){