+++ pastell-core/Field.class.php
@@ -9,8 +9,10 @@ class Field {
 	const LIBELLE_PROPERTIES_KEY = 'name'; /** Clé permettant de définir le libellé (lisible par un humain). Cette clé est improprement appelée "name" !*/
 	const INDEX_PROPERTIES_KEY = 'index'; /** Clé permettant d'indiquer si le champs doit-être indexé. La valeur de la clé est true ou false */
 	const VISIONNEUSE_PROPERTIES_KEY = 'visionneuse'; /** Clé permettant d'indiquer le nom d'une classe utilisé pour visualisé le fichier */
+	const VISIONNEUSE_NO_LINK = 'visionneuse-no-link';
 	const REQUIS = 'requis';
 	const DEFAULT = 'default';
+	const EDIT_ONLY = 'edit-only';
 	
 	private $fieldName;
 	private $properties;
@@ -167,6 +169,13 @@ class Field {
 	
 	public function getVisionneuse(){
 		return $this->getProperties(self::VISIONNEUSE_PROPERTIES_KEY);
+	}
+
+	public function displayLink(){
+		return ! $this->getProperties(self::VISIONNEUSE_NO_LINK);
+	}
+	public function isEditOnly(){
+		return $this->getProperties(self::EDIT_ONLY);
 	}
 	
 }