+++ template/DonneesFormulaireDetail.php
@@ -3,6 +3,7 @@
 /**
  * @var FieldData[] $fieldDataList
  * @var array $inject
+ * @var string $recuperation_fichier_url
  */
 
 $id_ce = $inject['id_ce'];
@@ -18,20 +19,25 @@ $id_e = $inject['id_e'];
 <?php endif; ?>
 	
 <table class='table table-striped'>
-<?php foreach($fieldDataList as $displayField): ?>
-		<tr>
+<?php foreach($fieldDataList as $num_field => $displayField):
+
+    if ($displayField->getField()->isEditOnly()){
+        continue;
+    }
+
+?>	    <tr>
 			<th class="w300">
 				<?php echo $displayField->getField()->getLibelle() ?>
 			</th>
 			<td>
 				<?php foreach($displayField->getValue() as $num => $value) :?>
 						<?php if ($displayField->isURL()) :?>
-							<a href='<?php echo $displayField->getURL($recuperation_fichier_url, $num,$id_e)?>'>
+							<a href='<?php echo $displayField->getURL($recuperation_fichier_url, $num,$id_e)?>' id="link_<?php echo $num_field?>">
 						<?php endif;?>
 							<?php if ($displayField->getField()->getType() == 'textarea') : ?>
 								<?php echo nl2br(get_hecho($value)); ?>
 							<?php else:?>
-							<?php hecho($value);?>
+                            <span><?php hecho($value);?></span>
 							<?php endif;?>
 							<br/>
 						<?php if($displayField->isURL()):?>
@@ -45,21 +51,20 @@ $id_e = $inject['id_e'];
                             class="btn btn-primary">
                             <i class="fa fa-download"></i>&nbsp;Télécharger tous les fichiers
                     </a>
-
                 <?php endif;?>
 
 				<?php if($displayField->getField()->getVisionneuse()):?>
-					<a class='visionneuse_link btn btn-primary' href='/DonneesFormulaire/visionneuse?id_e=<?php echo $id_e?>&id_d=<?php hecho($id_d)?>&id_ce=<?php hecho($id_ce); ?>&field=<?php hecho($displayField->getField()->getName()) ?>'>
+					<a id='visionneuse_link_<?php echo $num_field?>' class=' btn btn-primary' href='/DonneesFormulaire/visionneuse?id_e=<?php echo $id_e?>&id_d=<?php hecho($id_d)?>&id_ce=<?php hecho($id_ce); ?>&field=<?php hecho($displayField->getField()->getName()) ?>'>
                         <i class="fa fa-eye"></i>
                         &nbsp;Voir
                     </a>
 					<div class='visionneuse_result'></div>
 					<script>
 $(document).ready(function(){
-
+    let visionneuse_link = $('#visionneuse_link_<?php echo $num_field ?>');
 	$(".visionneuse_result").hide();
-	
-	$('.visionneuse_link').click(function(){
+
+    visionneuse_link.click(function(){
 		var link=$(this).attr('href');
 		var visionneuse_result = $(this).next(".visionneuse_result");
 		visionneuse_result.toggle();
@@ -71,9 +76,9 @@ $(document).ready(function(){
 			});
 		return false;
 	});
-
-
-	
+	<?php if (! $displayField->getField()->displayLink()) : ?>
+        $("#link_<?php echo $num_field ?>").hide();
+    <?php endif; ?>
 });
 					</script>