+++ CHANGELOG.md
@@ -39,6 +39,9 @@
 - Ajout de la classe StandardChoiceAction permettant d'utiliser des actions de connecteur type pour les choix externe
 - Ajout de la constante RGPD_PAGE_PATH permettant de mettre un fichier markdown contenant la politique vis à vis du RGPD #588
 - Api de supression de fichier #329
+- Ajout de la propriété `edit-only` afin de permettre l'affichage d'éléments uniquement en mode édition.
+- Ajout de la propriété `visionneuse-no-link` afin de permettre de supprimer le lien dans la présentation du dossier
+
 
 ## Corrections
 
@@ -77,6 +80,7 @@ Les fonctions suivantes sont dépréciées et seront retirées dans une prochain
 - Le script redis-flush-all.php est déprécié au profit de general-update.php
 - la classe Document au profit de la classe DocumentSQL
 - la table collectivite_fournisseur sera retirée dans la prochaine version
+- Pour modifier la typologie des actes, il faut passer par le champs externalData `type_piece` et plus par les champs `type_acte` et `type_pj`  
 
 # [2.0.13] - 2019-06-13